+++ api-processor/src/main/kotlin/io/bkbn/lerasium/api/processor/KtorProcessor.kt
@@ -1,6 +1,5 @@
 package io.bkbn.lerasium.api.processor
 
-import com.google.devtools.ksp.KspExperimental
 import com.google.devtools.ksp.processing.CodeGenerator
 import com.google.devtools.ksp.processing.KSPLogger
 import com.google.devtools.ksp.processing.Resolver
@@ -9,13 +8,11 @@ import com.google.devtools.ksp.symbol.KSAnnotated
 import com.google.devtools.ksp.symbol.KSClassDeclaration
 import com.google.devtools.ksp.validate
 import com.squareup.kotlinpoet.FileSpec
-import com.squareup.kotlinpoet.ksp.KotlinPoetKspPreview
 import com.squareup.kotlinpoet.ksp.writeTo
 import io.bkbn.lerasium.api.Api
 import io.bkbn.lerasium.utils.KotlinPoetUtils.BASE_API_PACKAGE_NAME
 import io.bkbn.lerasium.utils.LerasiumUtils.findParentDomain
 
-@OptIn(KotlinPoetKspPreview::class, KspExperimental::class)
 class KtorProcessor(
   private val codeGenerator: CodeGenerator,
   private val logger: KSPLogger,
@@ -39,24 +36,6 @@ class KtorProcessor(
       it.accept(ApiVisitor(fb, logger), Unit)
       val fs = fb.build()
       fs.writeTo(codeGenerator, false)
-    }
-
-    symbols.forEach {
-      val domain = it.findParentDomain()
-      val fb = FileSpec.builder(BASE_API_PACKAGE_NAME, domain.name.plus("ToC"))
-      it.accept(TocVisitor(fb, logger), Unit)
-      val fs = fb.build()
-      fs.writeTo(codeGenerator, false)
-    }
-
-    symbols.forEach {
-      val domain = it.findParentDomain()
-      val fb = FileSpec.builder(BASE_API_PACKAGE_NAME, domain.name.plus("Queries"))
-      it.accept(QueryModelVisitor(fb, logger), Unit)
-      val fs = fb.build()
-      if (fs.members.isNotEmpty()) {
-        fs.writeTo(codeGenerator, false)
-      }
     }
 
     return symbols.filterNot { it.validate() }.toList()